starrocks: implement nested type handling
This adjusts the syntax for json/map/array/jsonstring to match what Starrocks
expects. Along the way I fixed the type matching regexes, as these have to work
for nested types to be detected.

The most significant thing here is that the explicit json type detection which
the Clickhouse code does has not been reproduced. Starrocks does not provide a
way to introspect types inside a json blob, and so we rely on basic implicit
type coercion to make filtering work.

A more minor point is that I altered the prioritisation so that if a
JSON/Map/Array column is marked to be treated as a JSON String, it is still
evaluated as the native type. A JSON column is more efficiently handled this
way while Map and Array columns cannot be cast to JSON in the first place.